check if plugin is given in json file
this change came out of necessity due to the actions Jenkins took to hide their cas-plugin. A security hole was found in this plugin, so Jenkins removed the cas-plugin entry from it's [JSON file](https://updates.jenkins.io/current/update-center.actual.json). They also tried to hide cas-plugin.hpi, but I managed to download it by using a direct url to the version I wanted.

The pluginator was able to download cas-plugin.hpi but it couldn't find it's dependencies which threw a KeyError and exited the program.

This change allows the program to keep going but exit with a status of 1 if there was a problem. If a plugin isn't given in the [JSON file](https://updates.jenkins.io/current/update-center.actual.json), the culprate plugin is printed to STDOUT.